Server/BMAH: Convert currentBid and minIncrement to uint64. fixes an arithmetic overflow.

+++ src/server/game/BlackMarket/BlackMarketMgr.cpp
@@ -60,8 +60,8 @@ void BlackMarketAuction::SaveToDB(SQLTransaction& trans)
     stmt->setUInt32(1, GetTemplateId());
     stmt->setUInt32(2, GetStartTime());
     stmt->setUInt32(3, GetCurrentBidder());
-    stmt->setUInt32(4, GetCurrentBid());
-    stmt->setUInt32(5, GetMinIncrement());
+    stmt->setUInt64(4, GetCurrentBid());
+    stmt->setUInt64(5, GetMinIncrement());
     stmt->setUInt32(6, GetNumBids());
     trans->Append(stmt);
 }
@@ -77,8 +77,8 @@ void BlackMarketAuction::UpdateToDB(SQLTransaction& trans)
 {
     PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_UPD_BLACKMARKET_AUCTION);
     stmt->setUInt32(0, GetCurrentBidder());
-    stmt->setUInt32(1, GetCurrentBid());
-    stmt->setUInt32(2, GetMinIncrement());
+    stmt->setUInt64(1, GetCurrentBid());
+    stmt->setUInt64(2, GetMinIncrement());
     stmt->setUInt32(3, GetNumBids());
     stmt->setUInt32(4, GetAuctionId());
     trans->Append(stmt);
@@ -152,8 +152,8 @@ void BlackMarketMgr::LoadBlackMarketAuctions()
             uint32 templateId = fields[1].GetUInt32();
             uint32 startTime = fields[2].GetUInt32();
             uint32 currentBidder = fields[3].GetUInt32();
-            uint32 currentBid = fields[4].GetUInt32();
-            uint32 minIncrement = fields[5].GetUInt32();
+            uint64 currentBid = fields[4].GetUInt64();
+            uint64 minIncrement = fields[5].GetUInt64();
             uint32 numBids = fields[6].GetUInt32();
 
             BlackMarketAuction* auction = new BlackMarketAuction(auctionId, templateId, startTime, currentBidder, currentBid, minIncrement, numBids);
